Beginning to write code of Iterated Local Search

+++ src/types/solution.h
@@ -1,6 +1,6 @@
 typedef struct {
-    int *included;          // Array of indices of included items
-    double total_value;     // Solution value
-    double total_weight;    // Solution weight
-    double total_penalty;   // Penalty cost based on forfeit sets
+    int *included;          // Array of 0 or 1 indicating presence of item i
+    int total_value;     // Solution value
+    int total_weight;    // Solution weight
+    int total_penalty;   // Penalty cost based on forfeit sets
 } Solution;
